+++ BambooTracker/gui/pattern_editor/pattern_editor.hpp
@@ -88,6 +88,7 @@ public slots:
 	void onInterpolatePressed();
 	void onReversePressed();
 	void onReplaceInstrumentPressed();
+	void onFollowModeChanged();
 
 private:
 	Ui::PatternEditor *ui;