+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 ### Fixed
 - [#161] - Fix drawing error after loading song (thanks [@papiezak])
 - Fix groove start position (thanks [@maakmusic])
+- Fix some drawing errors in pattern editor and order list
 
 [#161]: https://github.com/rerrahkr/BambooTracker/issues/161